requirements.txt
attrs==17.2.0
autobahn==17.5.1
Automat==0.6.0
backports.ssl-match-hostname==3.5.0.1
certifi==2017.4.17
chardet==3.0.3
constantly==15.1.0
docker==2.3.0
docker-py==1.10.3
docker-pycreds==0.2.1
fastzbarlight==0.0.14
funcsigs==1.0.2
go-vncdriver==0.4.19
gym==0.7.4
h5py==2.7.0
idna==2.5
incremental==17.5.0
ipaddress==1.0.18
Keras==2.0.4
keras-rl==0.3.0
mock==2.0.0
numpy==1.13.0rc2
olefile==0.44
pbr==3.0.1
Pillow==4.1.1
protobuf==3.3.0
pyglet==1.2.4
PyYAML==3.12
requests==2.10.0
scipy==0.19.0
selenium==3.0.1
six==1.10.0
tensorflow==1.1.0
Theano==0.9.0
Twisted==17.1.0
txaio==2.7.1
ujson==1.35
universe==0.21.3
urllib3==1.21.1
websocket-client==0.40.0
Werkzeug==0.12.2
zope.interface==4.4.1